Move some functions into `clippy_utils::ty`
