+++ clippy_lints/src/methods/unnecessary_to_owned.rs
@@ -3,8 +3,11 @@ use super::unnecessary_iter_cloned::{self, is_into_iter};
 use clippy_utils::diagnostics::{span_lint_and_sugg, span_lint_and_then};
 use clippy_utils::msrvs::{self, Msrv};
 use clippy_utils::res::MaybeDef;
-use clippy_utils::source::{SpanExt, snippet, snippet_with_context};
-use clippy_utils::ty::{get_iterator_item_ty, implements_trait, is_copy, peel_and_count_ty_refs};
+use clippy_utils::source::{SpanExt, snippet};
+use clippy_utils::ty::{
+    get_callee_generic_args_and_args, get_iterator_item_ty, implements_trait, is_copy, is_to_string_on_string_like,
+    peel_and_count_ty_refs,
+};
 use clippy_utils::visitors::find_all_ret_expressions;
 use clippy_utils::{fn_def_id, get_parent_expr, is_expr_temporary_value, return_ty, sym};
 use rustc_errors::Applicability;
@@ -16,7 +19,7 @@ use rustc_lint::LateContext;
 use rustc_middle::mir::Mutability;
 use rustc_middle::ty::adjustment::{Adjust, Adjustment, DerefAdjustKind, OverloadedDeref};
 use rustc_middle::ty::{
-    self, ClauseKind, GenericArg, GenericArgKind, GenericArgsRef, ParamTy, ProjectionPredicate, TraitPredicate, Ty,
+    self, ClauseKind, GenericArg, GenericArgKind, ParamTy, ProjectionPredicate, TraitPredicate, Ty,
 };
 use rustc_span::Symbol;
 use rustc_trait_selection::traits::query::evaluate_obligation::InferCtxtExt as _;
@@ -131,10 +134,8 @@ fn check_addr_of_expr(
         && (*referent_ty != receiver_ty
             || (matches!(referent_ty.kind(), ty::Array(..)) && is_copy(cx, *referent_ty))
             || is_cow_into_owned(cx, method_name, method_parent_id))
+        && let Some(receiver_snippet) = receiver.span.get_text(cx)
     {
-        let mut applicability = Applicability::MachineApplicable;
-        let (receiver_snippet, _) = snippet_with_context(cx, receiver.span, expr.span.ctxt(), "..", &mut applicability);
-
         if receiver_ty == target_ty && n_target_refs >= n_receiver_refs {
             span_lint_and_sugg(
                 cx,
@@ -147,7 +148,7 @@ fn check_addr_of_expr(
                     "",
                     width = n_target_refs - n_receiver_refs
                 ),
-                applicability,
+                Applicability::MachineApplicable,
             );
             return true;
         }
@@ -167,8 +168,8 @@ fn check_addr_of_expr(
                     parent.span,
                     format!("unnecessary use of `{method_name}`"),
                     "use",
-                    receiver_snippet.to_string(),
-                    applicability,
+                    receiver_snippet.to_owned(),
+                    Applicability::MachineApplicable,
                 );
             } else {
                 span_lint_and_sugg(
@@ -178,7 +179,7 @@ fn check_addr_of_expr(
                     format!("unnecessary use of `{method_name}`"),
                     "remove this",
                     String::new(),
-                    applicability,
+                    Applicability::MachineApplicable,
                 );
             }
             return true;
@@ -193,7 +194,7 @@ fn check_addr_of_expr(
                 format!("unnecessary use of `{method_name}`"),
                 "use",
                 format!("{receiver_snippet}.as_ref()"),
-                applicability,
+                Applicability::MachineApplicable,
             );
             return true;
         }
@@ -398,6 +399,11 @@ fn check_other_call_arg<'tcx>(
         && let Some(as_ref_trait_id) = cx.tcx.get_diagnostic_item(sym::AsRef)
         && (trait_predicate.def_id() == deref_trait_id || trait_predicate.def_id() == as_ref_trait_id)
         && let receiver_ty = cx.typeck_results().expr_ty(receiver)
+        // Verify the output type contains the input type to be replaced.
+        // `needless_conversion_for_trait` cannot change the output type (see
+        // `clippy_utils::ty::replace_types`). Hence, this check ensures that `unnecessary_to_owned`
+        // and `needless_conversion_for_trait` do not flag the same code.
+        && fn_sig.output().contains(input)
         // We can't add an `&` when the trait is `Deref` because `Target = &T` won't match
         // `Target = T`.
         && let Some((n_refs, receiver_ty)) = if n_refs > 0 || is_copy(cx, receiver_ty) {
@@ -411,10 +417,8 @@ fn check_other_call_arg<'tcx>(
             None
         }
         && can_change_type(cx, maybe_arg, receiver_ty)
+        && let Some(receiver_snippet) = receiver.span.get_text(cx)
     {
-        let mut applicability = Applicability::MachineApplicable;
-        let (receiver_snippet, _) = snippet_with_context(cx, receiver.span, expr.span.ctxt(), "..", &mut applicability);
-
         span_lint_and_sugg(
             cx,
             UNNECESSARY_TO_OWNED,
@@ -422,7 +426,7 @@ fn check_other_call_arg<'tcx>(
             format!("unnecessary use of `{method_name}`"),
             "use",
             format!("{:&>n_refs$}{receiver_snippet}", ""),
-            applicability,
+            Applicability::MachineApplicable,
         );
         return true;
     }
@@ -441,33 +445,6 @@ fn skip_addr_of_ancestors<'tcx>(
         } else {
             return Some((parent, expr));
         }
-    }
-    None
-}
-
-/// Checks whether an expression is a function or method call and, if so, returns its `DefId`,
-/// `GenericArgs`, and arguments.
-fn get_callee_generic_args_and_args<'tcx>(
-    cx: &LateContext<'tcx>,
-    expr: &'tcx Expr<'tcx>,
-) -> Option<(
-    DefId,
-    GenericArgsRef<'tcx>,
-    Option<&'tcx Expr<'tcx>>,
-    &'tcx [Expr<'tcx>],
-)> {
-    if let ExprKind::Call(callee, args) = expr.kind
-        && let callee_ty = cx.typeck_results().expr_ty(callee)
-        && let ty::FnDef(callee_def_id, _) = callee_ty.kind()
-    {
-        let generic_args = cx.typeck_results().node_args(callee.hir_id);
-        return Some((*callee_def_id, generic_args, None, args));
-    }
-    if let ExprKind::MethodCall(_, recv, args, _) = expr.kind
-        && let Some(method_def_id) = cx.typeck_results().type_dependent_def_id(expr.hir_id)
-    {
-        let generic_args = cx.typeck_results().node_args(expr.hir_id);
-        return Some((method_def_id, generic_args, Some(recv), args));
     }
     None
 }
@@ -629,38 +606,12 @@ fn is_to_owned_like<'a>(
 ) -> bool {
     is_cow_into_owned(cx, method_name, method_parent_id)
         || (method_name != sym::to_string && is_clone_like(cx, method_name, method_parent_id))
-        || is_to_string_on_string_like(cx, call_expr, method_name, method_parent_id)
+        || is_to_string_on_string_like(cx, call_expr, method_parent_id)
 }
 
 /// Returns true if the named method is `Cow::into_owned`.
 fn is_cow_into_owned(cx: &LateContext<'_>, method_name: Symbol, method_parent_id: DefId) -> bool {
     method_name == sym::into_owned && method_parent_id.opt_impl_ty(cx).is_diag_item(cx, sym::Cow)
-}
-
-/// Returns true if the named method is `ToString::to_string` and it's called on a type that
-/// is string-like i.e. implements `AsRef<str>` or `Deref<Target = str>`.
-fn is_to_string_on_string_like<'a>(
-    cx: &LateContext<'_>,
-    call_expr: &'a Expr<'a>,
-    method_name: Symbol,
-    method_parent_id: DefId,
-) -> bool {
-    if method_name != sym::to_string || !method_parent_id.is_diag_item(cx, sym::ToString) {
-        return false;
-    }
-
-    if let Some(args) = cx.typeck_results().node_args_opt(call_expr.hir_id)
-        && let [generic_arg] = args.as_slice()
-        && let GenericArgKind::Type(ty) = generic_arg.kind()
-        && let Some(deref_trait_id) = cx.tcx.get_diagnostic_item(sym::Deref)
-        && let Some(as_ref_trait_id) = cx.tcx.get_diagnostic_item(sym::AsRef)
-        && (cx.get_associated_type(ty, deref_trait_id, sym::Target) == Some(cx.tcx.types.str_)
-            || implements_trait(cx, ty, as_ref_trait_id, &[cx.tcx.types.str_.into()]))
-    {
-        true
-    } else {
-        false
-    }
 }
 
 fn std_map_key<'tcx>(cx: &LateContext<'tcx>, ty: Ty<'tcx>) -> Option<Ty<'tcx>> {
